Refactor and improve services, controllers and variables naming
Renamed services, variables and controllers to improve readability and maintainability of the code. Removed two files `PasswordGenerator.java` and `RandomIDGenerator.java` as the functionality was moved to the `WishListService.java`. Enhanced Wish Service operation `addEmailToWish()` to add email to a specific wish. Also updated field references in HTML templates to match the renamed variables.