Introduce wish and wish-list deletion functionality
Added methods to allow for deletion of individual wishes and entire wish-lists from the application. Methods are included in the respective repositories, services and controllers. Also updated the corresponding HTML templates to include deletion buttons. Additionally, improvements made to the update method in WishService to use 'wishIntegerId' for more consistency and precision.

Am now done with all of the java coding.